closes #42. create PetType service impl

+++ pet-clinic-data/src/main/java/com/atech/repositories/PetTypeRepository.java
@@ -3,5 +3,5 @@ package com.atech.repositories;
 import com.atech.entity.PetType;
 import org.springframework.data.repository.CrudRepository;
 
-public interface PetTypeRepository extends CrudRepository<PetType, Integer> {
+public interface PetTypeRepository extends CrudRepository<PetType, Long> {
 }
